created a module to calculate today's date
